Set profiling calculation function as private
