+++ optimus/engines/base/basedataframe.py
@@ -5,7 +5,6 @@ import humanize
 import imgkit
 import jinja2
 import simplejson as json
-from glom import assign
 from tabulate import tabulate
 
 from optimus.engines.base.columns import *
@@ -791,209 +790,6 @@ class BaseDataFrame(ABC):
         # df = self.df
         df = self
         df.meta = {}
-        return df
-
-    def calculate_profile(self, cols="*", bins: int = MAX_BUCKETS, flush: bool = False, size=False):
-        """
-        Returns a new dataframe with the profile data in its added to the meta property
-        :param cols: "*", column name or list of column names to be processed.
-        :param bins:
-        :param flush:
-        :param size: get the dataframe size in memory. Use with caution this could be slow for big data frames.
-        :return:
-        """
-        _t = time.process_time()
-        profiler_time = {"hist": {}, "frequency": {}, "count_mismatch": {}}
-
-        df = self
-        meta = self.meta
-
-        if flush is False:
-            cols_to_profile = df._cols_to_profile(cols) or []
-        else:
-            cols_to_profile = parse_columns(df, cols) or []
-
-        profiler_data = Meta.get(meta, "profile")
-
-        is_cached = profiler_data is not None
-
-        if profiler_data is None:
-            profiler_data = {}
-        cols_data_types = None
-
-        profiler_time["beginning"] = {"elapsed_time": time.process_time() - _t}
-
-        if cols_to_profile or not is_cached or flush:
-
-            if flush:
-                meta = Meta.set(meta, "profile", {})
-                df.meta = meta
-
-            hist_cols = []
-            freq_cols = []
-
-            cols_data_types = {}
-            cols_to_infer = [*cols_to_profile]
-
-            for col_name in cols_to_profile:
-                col_data_type = Meta.get(df.meta, f"columns_data_types.{col_name}")
-
-                if col_data_type is not None:
-                    cols_data_types[col_name] = col_data_type
-                    cols_to_infer.remove(col_name)
-
-            if cols_to_infer:
-                cols_data_types = {**cols_data_types, **df.cols.infer_type(cols_to_infer, tidy=False)["infer_type"]}
-                cols_data_types = {col: cols_data_types[col] for col in cols_to_profile if col in cols_data_types}
-
-            _t = time.process_time()
-            mismatch = df.cols.quality(cols_data_types)
-            profiler_time["count_mismatch"] = {
-                "columns": cols_data_types, "elapsed_time": time.process_time() - _t}
-
-            # Get with columns are numerical and does not have mismatch so we can calculate the histogram
-            cols_properties = cols_data_types.items()
-            for col_name, properties in cols_properties:
-                if properties.get("data_type") in df.constants.NUMERIC_TYPES \
-                        and not properties.get("categorical", False):
-                    hist_cols.append(col_name)
-                else:
-                    freq_cols.append(col_name)
-
-            hist = None
-            freq = {}
-            sliced_freq = {}
-            count_uniques = None
-
-            if len(hist_cols):
-                _t = time.process_time()
-                hist = df.cols.hist(hist_cols, buckets=bins, compute=False)
-                profiler_time["hist"] = {
-                    "columns": hist_cols, "elapsed_time": time.process_time() - _t}
-
-            if len(freq_cols):
-                _t = time.process_time()
-                sliced_cols = []
-                non_sliced_cols = []
-
-                # Extract the columns with cells larger thatn
-                max_cell_length = getattr(df.meta, "max_cell_length", None)
-
-                if max_cell_length:
-                    for i, j in max_cell_length.items():
-                        if i in freq_cols:
-                            if j > 50:
-                                sliced_cols.append(i)
-                            else:
-                                non_sliced_cols.append(i)
-
-                else:
-                    non_sliced_cols = freq_cols
-
-                if len(non_sliced_cols) > 0:
-                    # print("non_sliced_cols",non_sliced_cols)
-                    freq = df.cols.frequency(
-                        non_sliced_cols, n=bins, count_uniques=True, compute=False)
-
-                if len(sliced_cols) > 0:
-                    # print("sliced_cols", sliced_cols)
-                    sliced_freq = df.cols.slice(sliced_cols, 0, 50).cols.frequency(sliced_cols, n=bins,
-                                                                                   count_uniques=True,
-                                                                                   compute=False)
-
-                profiler_time["frequency"] = {
-                    "columns": freq_cols, "elapsed_time": time.process_time() - _t}
-
-            def merge(_columns, _hist, _freq, _mismatch, _data_types, _count_uniques):
-                _c = {}
-
-                _hist = {} if _hist is None else _hist.get("hist", {})
-                _freq = {} if _freq is None else _freq.get("frequency", {})
-
-                for _col_name in _columns:
-                    _c[_col_name] = {
-                        "stats": _mismatch.get(_col_name, None),
-                        "data_type": _data_types.get(_col_name, None)
-                    }
-                    if _col_name in _freq:
-                        f = _freq[_col_name]
-                        _c[_col_name]["stats"]["frequency"] = f["values"]
-                        _c[_col_name]["stats"]["count_uniques"] = f["count_uniques"]
-
-                    elif _col_name in _hist:
-                        h = _hist[_col_name]
-                        _c[_col_name]["stats"]["hist"] = h
-
-                return {"columns": _c}
-
-            # Nulls
-            total_count_na = 0
-
-            data_types = df.cols.data_type("*", tidy=False)["data_type"]
-
-            hist, freq, sliced_freq, mismatch = self.functions.compute(
-                hist, freq, sliced_freq, mismatch)
-
-            freq = {**freq, **sliced_freq}
-
-            updated_columns = merge(
-                cols_to_profile, hist, freq, mismatch, data_types, count_uniques)
-
-            for col in list(profiler_data.get("columns", {}).keys()):
-                if col in updated_columns["columns"]:
-                    del profiler_data["columns"][col]
-
-            profiler_data = update_dict(profiler_data, updated_columns)
-
-            assign(profiler_data, "name", Meta.get(df.meta, "name"), dict)
-            assign(profiler_data, "file_name",
-                   Meta.get(df.meta, "file_name"), dict)
-
-            data_set_info = {'cols_count': df.cols.count(),
-                             'rows_count': df.rows.count(),
-                             }
-            if size is True:
-                data_set_info.update({'size': df.size(format="human")})
-
-            assign(profiler_data, "summary", data_set_info, dict)
-
-            data_types_list = list(set(df.cols.data_type("*", tidy=False)["data_type"].values()))
-
-            assign(profiler_data, "summary.data_types_list", data_types_list, dict)
-            assign(profiler_data, "summary.total_count_data_types",
-                   len(set([i for i in data_types.values()])), dict)
-            assign(profiler_data, "summary.missing_count", total_count_na, dict)
-
-            rows_count = df.rows.count()
-
-            if rows_count:
-                assign(profiler_data, "summary.p_missing", round(
-                    total_count_na / rows_count * 100, 2))
-            else:
-                assign(profiler_data, "summary.p_missing", None)
-
-        # _t = time.process_time()
-
-        all_columns_names = df.cols.names()
-
-        # meta = Meta.set(meta, "transformations", value={})
-
-        # Order columns
-        actual_columns = profiler_data["columns"]
-        profiler_data["columns"] = {key: actual_columns[key]
-                                    for key in all_columns_names if key in actual_columns}
-        meta = Meta.set(meta, "profile", profiler_data)
-
-        if cols_data_types is not None:
-            df.meta = meta
-            df = df.cols.set_data_type(cols_data_types, inferred=True)
-            meta = df.meta
-
-        # Reset Actions
-        meta = Meta.reset_actions(meta, parse_columns(df, cols or []))
-        df.meta = meta
-        profiler_time["end"] = {"elapsed_time": time.process_time() - _t}
-        # print(profiler_time)
         return df
 
     def graph(self) -> dict:
